+++ AlphaFS/Filesystem/DirectoryInfo Class/DirectoryInfo.ToString.cs
@@ -23,14 +23,11 @@ namespace Alphaleonis.Win32.Filesystem
 {
    partial class DirectoryInfo
    {
-      #region .NET
       /// <summary>Returns the original path that was passed by the user.</summary>
       /// <returns>A string that represents this object.</returns>
       public override string ToString()
       {
          return DisplayPath;
       }
-
-      #endregion // .NET
    }
 }