+++ AlphaFS/Filesystem/FileInfo Class/FileInfo.CopyToMoveTo.cs
@@ -628,7 +628,7 @@ namespace Alphaleonis.Win32.Filesystem
          FullPath = Path.GetRegularPathCore(destinationPathLp, GetFullPathOptions.None);
 
          OriginalPath = destinationPath;
-         DisplayPath = OriginalPath;
+         DisplayPath = Path.GetRegularPathCore(OriginalPath, GetFullPathOptions.None);
 
          _name = Path.GetFileName(destinationPathLp, true);