+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Version 2.1
 - Issue #60 : Remove all use of "Problematic" methods such as `DangerousAddRef` and `DangerousGetHandle()`.
 - Issue #160: `File.CreateSymbolicLink()` creates shortcut with no target. (martin-john-green)
 - Issue #162: `File.AppendAllLines()` concatenates content into one line. (pavelhritonenko)
+- Issue #169: `DirectoryInfo .ToString()` returns path with `\\UNC` prefix.
 - Issue #176: At `DirectoryInfo.GetFileSystemInfos()`, Long path prefix of GLOBALROOT path is missing. (diontools)
 - Isses #179: `Path.GetFileName()` with an empty string throws an exception. (brutaldev)
 - Issue #180: Network connects methods hangs in Windows service when credentials fail. (brutaldev)